// types.ts
export interface Customer {
  id: string;
  name: string;
  contact?: string;
  email?: string;
  address?: string;
  type?: 'B2B' | 'B2C'; 
  status?: 'ACTIVE' | 'INACTIVE' | 'VIP';
  notes?: string;
  joinDate?: string;
}